+++ Assets/01.Scripts/Character/CharacterDash.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7d63b8a67d1e58642874898e68af2102
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
