+++ Assets/01.Scripts/Character/CharacterJump.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e920d5e8cb5cafc47922259bcb322dcf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
