+++ Assets/01.Scripts/Character/Player/PlayerDash.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3a165c61ce3e6d544aef127d187255e6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
