+++ Assets/01.Scripts/Character/Player/PlayerJump.cs
@@ -1,0 +1,20 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerJump : CharacterJump
+{
+    private void Update()
+    {
+        InputJump();
+    }
+
+    // Input값을 바탕으로 점프하는 함수
+    private void InputJump()
+    {
+        if (InputManager.GetKeyDown(InputAction.Jump))
+        {
+            Jump(move.MoveStat.jumpForce);
+        }
+    }
+}
