+++ Assets/01.Scripts/Character/Player/PlayerJump.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 70ac393e02f7d8240bc242a9cd2b5bcd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
