+++ Assets/01.Scripts/Character/Player/PlayerMove.cs
@@ -1,0 +1,82 @@
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+/// <summary>
+/// 플레이어의 움직임을 담당하는 클래스
+/// </summary>
+public sealed class PlayerMove : CharacterMove
+{
+    private Vector3 forward = Vector3.zero;
+    private Vector3 right = Vector3.zero;
+    private Vector3 moveInput = Vector3.zero;
+
+    // 이거 나중에 상수로 빼든지 해서 어떻게 해야할 듯하다...
+    private readonly int isMoveHash = Animator.StringToHash("IsMove");
+    private PlayerDash dash;
+
+    protected override void Start()
+    {
+        animator = GetComponent<Animator>();
+        dash = GetComponent<PlayerDash>();
+
+        forward = GameManager.Instance.MainCam.transform.forward;
+        right = GameManager.Instance.MainCam.transform.right;
+
+        forward.y = 0f;
+        right.y = 0f;
+
+        base.Start();
+    }
+
+    private void Update()
+    {
+        InputMove();
+        InputRotate();
+    }
+
+    // Input값을 바탕으로 움직이는 함수
+    private void InputMove()
+    {
+        moveInput = Vector3.zero;
+
+        if (dash.IsDash) return;
+
+        if (InputManager.GetKey(InputAction.Up))
+            moveInput += forward;
+
+        if (InputManager.GetKey(InputAction.Down))
+            moveInput -= forward;
+
+        if (InputManager.GetKey(InputAction.Right))
+            moveInput += right;
+
+        if (InputManager.GetKey(InputAction.Left))
+            moveInput -= right;
+
+        Move(moveInput.normalized, moveStat.speed);
+    }
+
+
+    // 마우스쪽으로 바라보는 함수
+    private void InputRotate()
+    {
+        Camera cam = GameManager.Instance.MainCam;
+        float dist = Vector3.Distance(cam.transform.position, transform.position);
+        Ray cameraRay = cam.ScreenPointToRay(Input.mousePosition);
+
+        if (Physics.Raycast(cameraRay, cam.farClipPlane, Define.BOTTOM_LAYER))
+        {
+            Vector3 point = cameraRay.GetPoint(dist);
+            point.y = transform.position.y;
+
+            transform.LookAt(point);
+        }
+    }
+
+    protected override void OnMove(Vector3 velocity)
+    {
+        animator.SetBool(isMoveHash, velocity.sqrMagnitude > 0.1f);
+    }
+}
