+++ UserSettings/EditorUserSettings.asset
@@ -18,10 +18,13 @@ EditorUserSettings:
       value: 5404035552535a0b585a5d204676084446154f732a2a203578794964b3b1353e
       flags: 0
     RecentlyUsedSceneGuid-4:
-      value: 5109055e56065e5e0e0c5b7741200e44474e4f2c792b77322e284b61e6b4673e
+      value: 02570d565204500b080b592047740944174e4c79787c71617e7b4e63b6b5633b
       flags: 0
     RecentlyUsedSceneGuid-5:
-      value: 02570d565204500b080b592047740944174e4c79787c71617e7b4e63b6b5633b
+      value: 5153575e57510b0c0b0a5c2645710f4417164a7e2a7e76337e7b1835e4e4633e
+      flags: 0
+    RecentlyUsedSceneGuid-6:
+      value: 5109055e56065e5e0e0c5b7741200e44474e4f2c792b77322e284b61e6b4673e
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
